do not duplicate gatherer status creation
