+++ pkg/controller/periodic/periodic.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 	"sort"
-	"strings"
 	"time"
 
 	"k8s.io/apimachinery/pkg/api/errors"
@@ -20,6 +19,7 @@ import (
 	operatorv1client "github.com/openshift/client-go/operator/clientset/versioned/typed/operator/v1"
 	"github.com/openshift/insights-operator/pkg/anonymization"
 	"github.com/openshift/insights-operator/pkg/config/configobserver"
+	"github.com/openshift/insights-operator/pkg/controller/status"
 	"github.com/openshift/insights-operator/pkg/controllerstatus"
 	"github.com/openshift/insights-operator/pkg/gather"
 	"github.com/openshift/insights-operator/pkg/gatherers"
@@ -28,20 +28,6 @@ import (
 	batchv1 "k8s.io/api/batch/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-)
-
-const (
-	DataGatheredCondition = "DataGathered"
-	// NoDataGathered is a reason when there is no data gathered - e.g the resource is not in a cluster
-	NoDataGatheredReason = "NoData"
-	// Error is a reason when there is some error and no data gathered
-	GatherErrorReason = "GatherError"
-	// Panic is a reason when there is some error and no data gathered
-	GatherPanicReason = "GatherPanic"
-	// GatheredOK is a reason when data is gathered as expected
-	GatheredOKReason = "GatheredOK"
-	// GatheredWithError is a reason when data is gathered partially or with another error message
-	GatheredWithErrorReason = "GatheredWithError"
 )
 
 var (
@@ -406,7 +392,7 @@ func (c *Controller) updateOperatorStatusCR(allFunctionReports map[string]gather
 			continue
 		}
 
-		gs := createGathererStatus(&fr)
+		gs := status.CreateOperatorGathererStatus(&fr)
 		updatedOperatorCR.Status.GatherStatus.Gatherers = append(updatedOperatorCR.Status.GatherStatus.Gatherers, gs)
 	}
 
@@ -461,50 +447,6 @@ func (c *Controller) waitForJobCompletion(ctx context.Context, job *batchv1.Job)
 		// TODO check job conditions ?
 		return false, nil
 	}, ctx.Done())
-}
-
-func createGathererStatus(gfr *gather.GathererFunctionReport) v1.GathererStatus {
-	gs := v1.GathererStatus{
-		Name: gfr.FuncName,
-		LastGatherDuration: metav1.Duration{
-			// v.Duration is in milliseconds and we need nanoseconds
-			Duration: time.Duration(gfr.Duration * 1000000),
-		},
-	}
-	con := metav1.Condition{
-		Type:               DataGatheredCondition,
-		LastTransitionTime: metav1.Now(),
-		Status:             metav1.ConditionFalse,
-		Reason:             NoDataGatheredReason,
-	}
-
-	if gfr.Panic != nil {
-		con.Reason = GatherPanicReason
-		con.Message = gfr.Panic.(string)
-	}
-
-	if gfr.RecordsCount > 0 {
-		con.Status = metav1.ConditionTrue
-		con.Reason = GatheredOKReason
-		con.Message = fmt.Sprintf("Created %d records in the archive.", gfr.RecordsCount)
-
-		if len(gfr.Errors) > 0 {
-			con.Reason = GatheredWithErrorReason
-			con.Message = fmt.Sprintf("%s Error: %s", con.Message, strings.Join(gfr.Errors, ","))
-		}
-
-		gs.Conditions = append(gs.Conditions, con)
-		return gs
-	}
-
-	if len(gfr.Errors) > 0 {
-		con.Reason = GatherErrorReason
-		con.Message = strings.Join(gfr.Errors, ",")
-	}
-
-	gs.Conditions = append(gs.Conditions, con)
-
-	return gs
 }
 
 // PeriodicPrune runs periodically and deletes jobs (including the related pods) older
@@ -565,10 +507,6 @@ func (c *Controller) createNewDataGatherCR(disabledGatherers []string, dataPolic
 		},
 		Spec: insightsv1alpha1.DataGatherSpec{
 			DataPolicy: dataPolicy,
-		},
-		// TODO this shouldn't be here probably
-		Status: insightsv1alpha1.DataGatherStatus{
-			State: insightsv1alpha1.Pending,
 		},
 	}
 	for _, g := range disabledGatherers {
